requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
openai==1.12.0
python-docx==1.1.0
SQLAlchemy==2.0.27
flask-login==0.6.3
flask-migrate==4.0.5
python-jose==3.3.0
werkzeug==3.0.1
PyPDF2==3.0.1 